requirements.txt
flask==3.0.3
requests==2.32.3
sqlalchemy==2.0.36
alembic==1.14.0
textblob==0.18.0.post0
newsapi-python==0.2.7
flask-wtf==1.2.2
